dags/scripts/requirements.txt
psycopg2-binary==2.9.9
faker==18.13.0
faker_commerce==1.0.4